+++ tests/conftest.py
@@ -1,6 +1,4 @@
-import webbrowser
 from http.server import HTTPServer
-from unittest.mock import Mock
 
 import pytest
 
@@ -17,7 +15,7 @@ def disable_contract_checking():
 
 @pytest.fixture()
 def prevent_webbrowser_and_httpserver(mocker):
-    """Automatically mock webbrowser.open and HTTPServer.handle_request in all tests. this prevents any owser/server
+    """Automatically mock webbrowser.open and HTTPServer.handle_request in all tests. this prevents any browser/server
     code running when running Pytest to avoid CI timeouts, and unexpected browser popups."""
     mocker.patch("webbrowser.open", return_value=None)
     mocker.patch.object(HTTPServer, "handle_request", return_value=None)