+++ CHANGELOG.md
@@ -24,7 +24,6 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 🔧 Internal changes
 
-- Added mock `webbrowser.open` in tests to prevent browser tabs and HTTP requests during `python_ta.check_all()` executions.
 - Renamed `ExprWrapper` class to `Z3Parser`
 - Renamed `ExprWrapper` module to `z3_parser` and moved it to new directory `python_ta.z3`
 - Removed `node` attribute for `Z3Parser`
@@ -35,6 +34,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Checked for variable reassignment in `AugAssign` and `AnnAssign` node in parsing edge Z3 constraints
 - Rendered logically infeasible control flow graph edges in light grey
 - Modified `test_snapshot_to_json_sets_primitive` for Python 3.8 compatibility
+- Added unit tests for `one_iteration_checker`
+- Added mock `webbrowser.open` in tests to prevent browser tabs and HTTP requests during `python_ta.check_all()` executions.
 - Added `pytest-mock` as a development dependency
 
 ## [2.8.1] - 2024-08-19